Migrate fjs/ci/common to .f.mjs, split types into types.ts
Splits the type-level API (Os, Architecture, Image, Step, Job, Jobs,
GitHubAction, StepType, MetaStep) into a sibling types.ts, and
converts the runtime implementation (os, architecture, stepSchema,
jobSchema, jobsSchema, gitHubActionSchema, parseGitHubAction, uses,
install, test, toSteps, ubuntu, ubuntuArm) to JSDoc-typed
module.f.mjs under the stage-1 TypeScript-to-mjs migration.

Updates the nine dependents (fjs/ci itself, and its node/rust/deno/bun/nix
submodules) to import runtime values from module.f.mjs and types from
types.ts. Fixes five stale doc references in fjs/ci/todo and AGENTS.md.
