+++ fjs/ci/proof.f.ts
@@ -4,7 +4,8 @@ import { nodeNixJobs } from './node/module.f.ts'
 import { coverageInclude } from './deno/module.f.ts'
 import { utf8, utf8ToString } from '../text/module.f.mjs'
 import { empty as emptyVec } from '../types/bit_vec/module.f.mjs'
-import { type MetaStep, type Os, test, ubuntu, type GitHubAction, parseGitHubAction } from './common/module.f.ts'
+import { test, ubuntu, parseGitHubAction } from './common/module.f.mjs'
+import type { MetaStep, Os, GitHubAction } from './common/types.ts'
 import { assert, assertEq } from '../asserts/module.f.mjs'
 import type { State } from '../effects/node/virtual/module.f.ts'
 import { emptyState, virtual, type Dir } from '../effects/node/virtual/module.f.ts'
