+++ fjs/ci/todo/170.md
@@ -48,7 +48,7 @@ test-command strings:
 ### Proposed abstraction
 
 A per-runtime data record consumed by one factory in
-`fjs/ci/common/module.f.ts`:
+`fjs/ci/common/module.f.mjs`:
 
 ```ts
 export type ToolRecipe = {
@@ -94,7 +94,7 @@ install may vary per `(Os, Architecture)`; deno and node pass a fixed
 
 ### Related
 
-- `fjs/ci/common/module.f.ts` already centralizes
+- `fjs/ci/common/module.f.mjs` already centralizes
   `install`/`test`/`uses`/`toSteps`; `toolSteps` is the next
   composition up the same ladder.
 - [66h-ci-npm-global-install.md](./66h-ci-npm-global-install.md) — the
