+++ fjs/ci/todo/175.md
@@ -26,7 +26,7 @@ install({ uses: 'wasmerio/setup-wasmer@v3.1', with: { version: `v${wasmer}` } })
 
 ### Proposed abstraction
 
-A small factory in `fjs/ci/common/module.f.ts` that captures "a setup action
+A small factory in `fjs/ci/common/module.f.mjs` that captures "a setup action
 parameterized by one pinned version":
 
 ```ts
